refactor: [FE] 게임페이지 소켓 코드 훅으로 분리
