+++ frontend/src/atoms/game.ts
@@ -130,3 +130,8 @@ export const canvasSelectedColorState = atom<string>({
     key: 'canvasSelectedColorState',
     default: PEN_DEFAULT_COLOR,
 });
+
+export const isRoundTimeoutState = atom<boolean>({
+    key: 'isRoundTimeoutState',
+    default: false,
+});
