+++ frontend/src/pages/Game.tsx
@@ -1,19 +1,14 @@
 import { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import { ScaledDiv, ScaledSection } from '@styles/styled';
-import { onCompleteGame, onCountSubmittedQuiz } from '@game/NetworkServiceUtils';
-import { useSetRecoilState, useRecoilState, useRecoilValue } from 'recoil';
-import { gameResultIdState, gameResultState } from '@atoms/result';
-import { resetModalOpenState, submittedQuizReplyCountState, userListState } from '@atoms/game';
+import { useRecoilValue } from 'recoil';
+import { resetModalOpenState } from '@atoms/game';
 import GameUsers from '@components/GameUsers';
 import MicButton from '@components/MicButton';
 import CameraButton from '@components/CameraButton';
 import SmallLogo from '@assets/zellodraw-logo.png';
 import GameSketchbook from '@components/GameSketchbook';
 import ResultSketchbook from '@components/resultSketchbook/ResultSketchbook';
-import { networkServiceInstance as NetworkService } from '../services/socketService';
-import { JoinLobbyReEmitRequest } from '@backend/core/user.dto';
-import { userState } from '@atoms/user';
 import useBeforeReload from '@hooks/useBeforeReload';
 import CountDown from '@components/CountDown';
 import { AnimatePresence } from 'framer-motion';
@@ -21,57 +16,20 @@ import ResetModal from '@components/ResetModal';
 import SoundControlButton from '@components/SoundControlButton';
 import countdownSound from '@assets/sounds/countdown.mp3';
 import useSoundEffect from '@hooks/useSoundEffect';
+import useGameSocket from '@hooks/socket/useGameSocket';
 
 function Game() {
-    const [user, setUser] = useRecoilState(userState);
-    const setGameResult = useSetRecoilState(gameResultState);
-    const setGameResultId = useSetRecoilState(gameResultIdState);
-    const setUserList = useSetRecoilState(userListState);
-    const setSubmittedQuizReplyCount = useSetRecoilState(submittedQuizReplyCountState);
     const [isCompleteGame, setIsCompleteGame] = useState(false);
     const [isStarted, setIsStarted] = useState(false);
     const resetModalOpen = useRecoilValue(resetModalOpenState);
     const { playSoundEffect } = useSoundEffect();
     useBeforeReload();
+    const { emitSubmitQuizReply } = useGameSocket(setIsCompleteGame);
 
     useEffect(() => {
         playSoundEffect(countdownSound);
         setTimeout(() => setIsStarted(true), 2500);
     }, []);
-
-    useEffect(() => {
-        onCountSubmittedQuiz(setSubmittedQuizReplyCount);
-        onCompleteGame(setGameResultId, setGameResult, setIsCompleteGame);
-
-        NetworkService.on('update-user-stream', (payload) => {
-            setUserList((prev) =>
-                prev.map((user) => {
-                    const prevUserValue = { ...user };
-                    if (payload.socketId === user.sid) {
-                        prevUserValue.audio = payload.audio;
-                        prevUserValue.video = payload.video;
-                    }
-                    return prevUserValue;
-                }),
-            );
-        });
-
-        NetworkService.on('leave-game', (user: JoinLobbyReEmitRequest) => {
-            setUserList((prev) =>
-                prev.filter((participant) => participant.userName !== user.userName),
-            );
-        });
-
-        NetworkService.on('succeed-host', () => {
-            setUser({ ...user, isHost: true });
-        });
-
-        return () => {
-            NetworkService.off('update-user-stream');
-            NetworkService.off('leave-game');
-            NetworkService.off('succeed-host');
-        };
-    }, [isStarted]);
 
     return (
         <>
@@ -82,7 +40,7 @@ function Game() {
                     {isCompleteGame ? (
                         <ResultSketchbook isForShareResult={false} />
                     ) : (
-                        <GameSketchbook />
+                        <GameSketchbook emitSubmitQuizReply={emitSubmitQuizReply} />
                     )}
                 </SketchbookSection>
                 {resetModalOpen && <ResetModal />}
